Code Revamp
I Was having trouble with my code having errors so i went back and
re-wrote the code top to bottom to ensure correct syntax throughout.  I
also added the time time variable setting the injected current to 0;
therefore, testing the steady state condition.  The dice was found to
work in this situation